DBZ-537 Move OffsetCommitPolicy to SPI
